requirements.txt
python-telegram-bot
gspread
oauth2client
google-generativeai
g4f>=0.3.5.7
pandas
openpyxl
yt-dlp>=2025.3.31
ffmpeg-python
setuptools-rust
apscheduler
xlsxwriter
